+++ dotnet/src/AuthTokenHandling.TestApp/Program.cs
@@ -1,8 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+
+[assembly: AssemblyMetadata("SourceContext", "AuthTokenHandling.TestApp")]
 
 namespace AuthTokenHandling.TestApp {