+++ dotnet/src/AuthTokenHandling.WinFormsClient/Issuing/OAuthInteractiveAuthenticator.cs
@@ -7,9 +7,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Reflection;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+
+[assembly: AssemblyMetadata("SourceContext", "AuthTokenHandling")]
 
 namespace Security.AccessTokenHandling {